# Profile screen performance tweak - Adjust initial num to render based on header height