--- a/src/lib/hooks/useInitialNumToRender.ts
+++ b/src/lib/hooks/useInitialNumToRender.ts
@@ -1,5 +1,7 @@
 import {useWindowDimensions} from 'react-native'
-import {useBottomTabBarHeight} from '@react-navigation/bottom-tabs'
+import {useSafeAreaInsets} from 'react-native-safe-area-context'
+
+import {useBottomBarOffset} from 'lib/hooks/useBottomBarOffset'
 
 const MIN_POST_HEIGHT = 100
 
@@ -8,7 +10,10 @@ export function useInitialNumToRender({
   screenHeightOffset = 0,
 }: {minItemHeight?: number; screenHeightOffset?: number} = {}) {
   const {height: screenHeight} = useWindowDimensions()
-  const bottomBarHeight = useBottomTabBarHeight()
-  const finalHeight = screenHeight - screenHeightOffset - bottomBarHeight
+  const {top: topInset} = useSafeAreaInsets()
+  const bottomBarHeight = useBottomBarOffset()
+
+  const finalHeight =
+    screenHeight - screenHeightOffset - topInset - bottomBarHeight
   return Math.floor(finalHeight / minItemHeight) + 1
 }