## Why

Whenever we first open a screen, we always want to initially render as few items as are necessary to fill the screen _and only fill the screen_. This results in faster tab switches and navigation pushes.

Right now, the profile tab will attempt to render around 9 posts initially on my profile, which is what would normally be needed to fill the screen. However, that number does not take into account the height of the profile header. Because the header often takes up half or more of the screen, we end up rendering a ton of extra posts that don't need to be there.

If we account for the header height, then the number of posts that are initially rendered when visiting my profile drops to four, cutting the initial render time of the screen by around half.

## Test Plan

Profile before and after

| Before | After |
|--------|--------|
| ![Screenshot 2024-08-27 at 5 45 51 PM](https://github.com/user-attachments/assets/d16be369-7bb2-4ba6-8a22-51a17aaefb17) | ![Screenshot 2024-08-27 at 5 49 29 PM](https://github.com/user-attachments/assets/560f2232-e77c-454e-aa4f-6fcad1fb755f) |